fixing unit tests

+++ marshall_webapp/views/tests/test_views_resources_transients_element.py
@@ -14,31 +14,34 @@ packageDirectory = utKit("").get_project_root()
 # settingsFile = packageDirectory + "/test_settings.yaml"
 settingsFile = home + "/git_repos/_misc_/settings/marshall/test_settings.yaml"
 
-su = tools(
-    arguments={"settingsFile": settingsFile},
-    docString=__doc__,
-    logLevel="DEBUG",
-    options_first=False,
-    projectName=None,
-    defaultSettingsFile=False
-)
-arguments, settings, log, dbConn = su.setup()
+exists = os.path.exists(settingsFile)
+if exists:
+    su = tools(
+        arguments={"settingsFile": settingsFile},
+        docString=__doc__,
+        logLevel="DEBUG",
+        options_first=False,
+        projectName=None,
+        defaultSettingsFile=False
+    )
+    arguments, settings, log, dbConn = su.setup()
 
-# SETUP PATHS TO COMMON DIRECTORIES FOR TEST DATA
-moduleDirectory = os.path.dirname(__file__)
-pathToInputDir = moduleDirectory + "/input/"
-pathToOutputDir = moduleDirectory + "/output/"
+    # SETUP PATHS TO COMMON DIRECTORIES FOR TEST DATA
+    moduleDirectory = os.path.dirname(__file__)
+    pathToInputDir = moduleDirectory + "/input/"
+    pathToOutputDir = moduleDirectory + "/output/"
 
-try:
-    shutil.rmtree(pathToOutputDir)
-except:
-    pass
-# COPY INPUT TO OUTPUT DIR
-shutil.copytree(pathToInputDir, pathToOutputDir)
+    try:
+        shutil.rmtree(pathToOutputDir)
+    except:
+        pass
+    # COPY INPUT TO OUTPUT DIR
+    shutil.copytree(pathToInputDir, pathToOutputDir)
 
-# Recursively create missing directories
-if not os.path.exists(pathToOutputDir):
-    os.makedirs(pathToOutputDir)
+    # Recursively create missing directories
+    if not os.path.exists(pathToOutputDir):
+        os.makedirs(pathToOutputDir)
+
 
 class test_views_resources_transients_element(BaseTest):
